Add support for encoding packets.
When encoding, you must specify a code (e.g. "Access-Request"). If you want
to send a password, you must give the RADIUS secret as well. If you don't
specify an identifier (sequence number), it will randomly choose one. If you
don't specify a request authenticator, it will generate a random one for you.
And, of course, you can specify the attributes you want to put into your
message. If there is no dictionary for the attribute you want to send, you
must proved a Buffer for the value (since it won't know how to encode it).