Handle repeated attributes.
Some attributes can be repeated (with different values) multiple times
in a single packet. Now if we see an attribute again while decoding, we
will turn the "attributes" value into an array containing all the values.
Unfortunately the dictionary files do not encode which fields can show
up multiple times, so the developer needs to check if a repeatable
attribute's value is an array or a single value.