Accept attributes as hash.
For simple packets you can pass the attributes as an array or
as a hash. This will only work if you do not have vendor specific attributes
and the order of the attributes does not matter for your packet.

If you have nested attributes this will throw an exception.

It is still recommended that you use the array style as that will work in
all cases.