Fix encoding after using decode_without_secret.
The way the library passes args down is really stupid. It sticks
various args on "this" (the "radius" package object) so it doesn't
have to pass the args around as function parameters. This was causing
and issue because the encode_without_secret method was setting
this.no_secret = true (which indicates to _crypt_field that it shouldn't
try to decrypt anything), but encode wasn't unsetting that, it was
skipping the password encrypt and crashing later because of it.

Double fix it by having encode unset no_password, and by making
decrypt_field consume no_secret rather than _crypt_field (so it
can't affect encrypt_field anymore).